Move code in src/functions/wavelet_ast/mod.rs to module files
The PR moves the code in src/functions/wavelet_ast/mod.rs to:

- data.rs (WaveletSpec, parse_discrete_wavelet),
- continuous.rs (ContinuousWaveletSpec, parse_continuous_wavelet)
- filters.rs (wavelet_filters, wavelet_filter_coefficients_ast)
- transforms.rs (spec_to_expr, now private)
