+++ src/functions/wavelet_ast/transforms.rs
@@ -10,6 +10,7 @@
 //! - Decimated coefficient arrays at the next level have length
 //!   ceil((n + fl - 2)/2) where fl is the analysis filter length.
 
+use super::data::WaveletSpec;
 use super::filters::{Filter, WaveletFilters, highpass_from};
 use crate::syntax::Expr;
 
@@ -955,10 +956,47 @@ pub fn basis_index(kind: TransformKind, r: usize, rank: usize) -> Vec<Vec<u8>> {
 /// LiftingFilterData[…] placeholder object produced by
 /// WaveletFilterCoefficients[wave, "LiftingFilter"]: stores the wavelet so
 /// filter properties and LiftingWaveletTransform can consume it.
-pub fn lifting_filter_data(spec: &super::WaveletSpec) -> Expr {
-  let wavelet = super::spec_to_expr(spec);
+pub fn lifting_filter_data(spec: &WaveletSpec) -> Expr {
+  let wavelet = spec_to_expr(spec);
   Expr::FunctionCall {
     name: "LiftingFilterData".to_string(),
     args: vec![wavelet].into(),
+  }
+}
+
+/// Canonical expression for a discrete wavelet spec.
+fn spec_to_expr(spec: &WaveletSpec) -> Expr {
+  let call = |name: &str, args: Vec<Expr>| Expr::FunctionCall {
+    name: name.to_string(),
+    args: args.into(),
+  };
+  let int = |n: i128| Expr::Integer(n);
+  match spec {
+    WaveletSpec::Haar => call("HaarWavelet", vec![]),
+    WaveletSpec::Daubechies(n) => {
+      call("DaubechiesWavelet", vec![int(*n as i128)])
+    }
+    WaveletSpec::Symlet(n) => call("SymletWavelet", vec![int(*n as i128)]),
+    WaveletSpec::Coiflet(n) => call("CoifletWavelet", vec![int(*n as i128)]),
+    WaveletSpec::BattleLemarie(n, lim) => call(
+      "BattleLemarieWavelet",
+      vec![int(*n as i128), Expr::Real(*lim)],
+    ),
+    WaveletSpec::BiorthogonalSpline(n, m) => call(
+      "BiorthogonalSplineWavelet",
+      vec![int(*n as i128), int(*m as i128)],
+    ),
+    WaveletSpec::ReverseBiorthogonalSpline(n, m) => call(
+      "ReverseBiorthogonalSplineWavelet",
+      vec![int(*n as i128), int(*m as i128)],
+    ),
+    WaveletSpec::Cdf(lossy) => call(
+      "CDFWavelet",
+      vec![Expr::String(if *lossy { "9/7" } else { "5/3" }.to_string())],
+    ),
+    WaveletSpec::Meyer(n, lim) => {
+      call("MeyerWavelet", vec![int(*n as i128), Expr::Real(*lim)])
+    }
+    WaveletSpec::Shannon(lim) => call("ShannonWavelet", vec![Expr::Real(*lim)]),
   }
 }
